resolved: demote only the fallback servers in the global scope
This softens the behavior introduced in c83cd8da4916, and simultaneously
expands the role of fallback servers so they are applied not only when
there are no dns servers configured, but when all the configured dns
servers are configured only for non-default-route links.

The intent is that the global FallbackDNS (instead of DNS) can now be
used in conjunction with the per-link dns, providing a fallback behavior
without introducing a scope overlap.

Fixes: c83cd8da4916 (resolved: demote the global unicast scope)